Custom render-function for routes
Route-functions can have a custom render-function, which allows to override the default render-engine
for that route. Render-functions are expected to handle all rendering theirself, there is no direct
relation to the specified render-engine, although if could be accessed via the site-instance or by
importing the module, if required.

A use-case would be custom content-type rendering, i.e. responses for POST-requests.